refactor: split linter constants
