+++ src/linter/rules/invalid-change-version.mjs
@@ -1,4 +1,4 @@
-import { LINT_MESSAGES } from '../../constants.mjs';
+import { LINT_MESSAGES } from '../constants.mjs';
 import { valid } from 'semver';
 
 /**
